Redo source distribution targets to explicitly change line endings.
'abcl.source.{tar,zip}' now explicit changes the EOL for most files to
'lf' (UNIX) for the tar, and to 'crf' (DOS) for the zip.  The current
exception is that the 'abcl.in' script always gets 'lf' EOL, and the
'abcl.bat.in' always get 'crlf' EOL.

The 'abcl.source.eol' property has been removed.



git-svn-id: http://abcl.org/svn@12127 1c010e3e-69d0-11dd-93a8-456734b0d56f