Ant-based build process now records FASL source locations correctly.
Now we no longer copy the Lisp system sources to the build directory,
instead directly referring to the actual source location.  As a
result, the FASLs now correctly record the location of the system
source files.  This makes using SLIME to edit system source a lot
saner.

Specifying the JVM property 'abcl.home' now overrides the dynamic
lookup for 'boot.lisp' on the classpath of org.armedbear.lisp.Lisp for
setting the EXT::*LISP-HOME* property.



git-svn-id: http://abcl.org/svn@12291 1c010e3e-69d0-11dd-93a8-456734b0d56f