ABCL packed FASLS no longer have to end in ".abcl".
The requirement that packed FASLS had to end in ".abcl" was introduced
in the recent Pathname code as a optimization to avoid opening/closing
an input stream on the assumption that renaming FASLs would not be
expected to work.  But COMPILE-FILE has an :OUTPUT-FILE arg that
certainly can be used, so we relax this restriction.

Found by dmiles in the ANSI test suite.




git-svn-id: http://abcl.org/svn@12426 1c010e3e-69d0-11dd-93a8-456734b0d56f