Implement HTTP HEAD Last-Modified checking for ZipCache objects.
Since it seems that no Sun-derived JVM ever invalidates its cache of
URLConnection objects, we have to check for modification "manually".
This implementation is "better than nothing", but not expected to be
especially robust.  Most notably, this implementation does not attempt
to use http proxies if they have been specified to the JVM by the
'http.proxy' system property.


git-svn-id: http://abcl.org/svn@12504 1c010e3e-69d0-11dd-93a8-456734b0d56f