ANSI test database can now contain multiple test results per version.
We change the syntax of the ANSI test results database to allow the
specification of a unique identifier plus other optional identifying
information by allowing keyword/value pairs.  The keyword :ID
specifies the identifier, which should be a symbol.  Other arbitrary
keywords are allowed which specify additional information to be
associated with the symbol specified by :ID in the *ID* hashtable.
Not every test failure entry needs to specify this information.  In
case of duplicates, the last entry wins.  Suggested other keywords are
:JVM to specify the Java virtual machine, and :UNAME to specify the
operating system/hardware combination in a GNU autoconf-like string.
See the comments at the beginning of 'parse-ansi-errors.lisp' for more
details.

The utility has been packaged in ABCL.ANSI.TEST, showing up in the
ANSI-COMPILED and ANSI-INTERPRETED ASDF systems loadable from
'abcl.asd'.

A database of failures has been included in 'ansi-test-failures'.  It
is intended that other developers entrich this database with their own
test results.



git-svn-id: http://abcl.org/svn@12509 1c010e3e-69d0-11dd-93a8-456734b0d56f