Add interfaces, superclasses and member classes for JAVA-OBJECT inspector.
Make the display of a java Class under SLIME (and via the built-in
inspector) a bit more useful by displaying a list of all superclasses
in "Superclasses", a list of all implemented interfaces in
"Interfaces", and a list of all member classes in "Member classes".

In general the information returned by the getParts() inspector
contract for java classes could be made quite a bit more useful by
adding the output of more methods such as getAnnotations() and
getCannonicalName() but for those not using the SLIME inspector there
may actually be too much information presented.

git-svn-id: http://abcl.org/svn@13278 1c010e3e-69d0-11dd-93a8-456734b0d56f