Strip VERSION from the source location output by COMPILE-FILE.
ABCL currently outputs non-readable--by the Lisp reader--namestrings
for any PATHNAME with a non-NIL VERSION.  With the recent fixes to how
ABCL deals with VERSION components by defaulting to the ANSI specified
:NEWEST in MERGE-PATHNAMES, this breaks ASDF compiliation which tends
to acculumlate such PATHNAMES in its labryinth of cross-implementation
TRUENAMIZE* and LISPIZE-PATHNAME invocations.  Since VERSION doesn't
really have an effect on the filesystems that ABCL deals with we just
silently nip it to NIL before forming our FASL source location.

git-svn-id: http://abcl.org/svn@13290 1c010e3e-69d0-11dd-93a8-456734b0d56f