Fix #183: move threads-jss.lisp out of system source.
'threads-jss.lisp' provides a rudimentary implementation of a server
framework using the java.util.concurrent abstractions.  Ripped out of
another project, the code uses the JSS syntax for brevity making it
more or less impossible to actually compile as system source as it
depends on the JSS contrib.  We move it to the tools directory until
we can rewrite the use of primitives to use the lower-level Java FFI.

git-svn-id: http://abcl.org/svn@13703 1c010e3e-69d0-11dd-93a8-456734b0d56f