Fix for maven-3.0.4.
Part of the process of cleaning up ABCL-ASDF to work with more Maven
versions.

Seems to be failing for OS X Maven 3.0.4, so need to do a round of testing.

Add system definition of ABCL-ASDF-TEST which becomes the behavior for
invoking ASDF:TEST-SYSTEM on ABCL-ASDF.  Really update to version
"0.5.0".

Correct misspellings.

TODO:  figure out a list of repos to iteratively search.

TODO:  overload the system defintion version mechanism to allow
non-integer values, or just use the IRI class?

TODO:  be more verbose on searching/downloading Maven artifacts as it
can be quite slow for large amounts of artifacts.

git-svn-id: http://abcl.org/svn@13804 1c010e3e-69d0-11dd-93a8-456734b0d56f