abcl-asdf: Clarify the version setting code for MVN compoments.
Fix #204, but considerably beyond the patch the Cyrus Harmon submitted.

Drastically simplified the ASDF:MAYBE-PARSE-MVN function to take
advantage of the fact that we are in the ASDF namespace.  At least the
function now makes sense

Parts of ASDF *reallY* want ASDF:VERSION to be a triple of intergers,
and never anything more, which is part of the reason for the
shennigans here.

git-svn-id: http://abcl.org/svn@13900 1c010e3e-69d0-11dd-93a8-456734b0d56f