compiler: don't signal conditions for fasl verification error and muffle diagnostics by default.
HEADS UP: problems seem to exist ANSI tests, which triggers the
attempt to load the fasl classfile to verify its integrity.

Don't signal problems just yet, until satisfied that the correct
diagnostic messages are being triggered.

Refactored diagnostics interface to use a new SYS::DIAG macro whose output is
directed to the value of SYS:*COMPILER-DIAGNOSTIC*.  This should be
reconsidered in view of all the diagnostic frameworks when I
understand how they are to be used

git-svn-id: http://abcl.org/svn@13915 1c010e3e-69d0-11dd-93a8-456734b0d56f