quicklisp-abcl:  use http://beta.quicklisp.org/quicklisp.lisp as the bootstrap URI.
THe JVM seems confused with the AMS CloudFront certificate wildcards
as to whether 'https://beta.quicklisp.org' is valid or not, so switch to using "https".

N.b. ASDF loading of QUICKLISP-ABCL currently doesn't trigger the
load-op, probably because we have no declared artifacts, so this
contrib cannot be considered as "working" yet.  But if the
ASDF:PERFORM *were* to get called, things should work.

git-svn-id: http://abcl.org/svn@14223 1c010e3e-69d0-11dd-93a8-456734b0d56f