abcl-asdf:  Enable bypassing of loading from network if a given class already exists.
An ASDF:MVN component can now optionally specify a CLASSNAME that if
able to be found in the current jvm process, inhbits further loading
from the network.

An ASDF:MVN component may also optionally specify an ALTERNATE-URI
that will be added to the current jvm classpath if Maven cannot be
invoked.  Most jvm implementations won't access such jar archives from
the network by default, so this is an intermediate step before
actually writing the code to download the jar to the local filesystem
to then be added.

Refresh documentation in README.markdown.

git-svn-id: http://abcl.org/svn@14236 1c010e3e-69d0-11dd-93a8-456734b0d56f