aio: use jar manifest to declare location of packaged entities
A Jar manifest may have sections which we use to declare locations
within the Jar for packaged entities.  The key 'name' denotes the
relative path within the jar.  The key 'Implementation-Title' uniquely
defines an id for the packaged entity.  We define 'org.abcl-contrib'
to be the key that identifies the location of ABCL-CONTRIB, and we
modify the search strategy defined by SYS:FIND-CONTRIB and friends
accordingly.

<http://docs.oracle.com/javase/8/docs/technotes/guides/jar/jar.html>

git-svn-id: http://abcl.org/svn@14908 1c010e3e-69d0-11dd-93a8-456734b0d56f